+++ custom_components/tahoma/sensor.py
@@ -1,16 +1,26 @@
-"""Support for Tahoma sensors."""
+"""Support for TaHoma sensors."""
 from datetime import timedelta
 import logging
+from typing import Optional
 
-from homeassistant.const import ATTR_BATTERY_LEVEL, TEMP_CELSIUS, UNIT_PERCENTAGE
+from homeassistant.const import (
+    ATTR_BATTERY_LEVEL,
+    CONCENTRATION_PARTS_PER_MILLION,
+    POWER_WATT,
+    TEMP_CELSIUS,
+    UNIT_PERCENTAGE,
+)
 from homeassistant.helpers.entity import Entity
 
 from .const import (
-    DOMAIN,
-    TAHOMA_TYPES,
-    CORE_RELATIVE_HUMIDITY_STATE,
+    CORE_CO2_CONCENTRATION_STATE,
+    CORE_ELECTRIC_POWER_CONSUMPTION_STATE,
     CORE_LUMINANCE_STATE,
+    CORE_RELATIVE_HUMIDITY_STATE,
     CORE_TEMPERATURE_STATE,
+    DOMAIN,
+    TAHOMA_SENSOR_DEVICE_CLASSES,
+    TAHOMA_TYPES,
 )
 from .tahoma_device import TahomaDevice
 
@@ -20,7 +30,7 @@ SCAN_INTERVAL = timedelta(seconds=60)
 
 
 async def async_setup_entry(hass, entry, async_add_entities):
-    """Set up the Tahoma sensors from a config entry."""
+    """Set up the TaHoma sensors from a config entry."""
 
     data = hass.data[DOMAIN][entry.entry_id]
 
@@ -35,7 +45,7 @@ async def async_setup_entry(hass, entry, async_add_entities):
 
 
 class TahomaSensor(TahomaDevice, Entity):
-    """Representation of a Tahoma Sensor."""
+    """Representation of a TaHoma Sensor."""
 
     def __init__(self, tahoma_device, controller):
         """Initialize the sensor."""
@@ -53,7 +63,8 @@ class TahomaSensor(TahomaDevice, Entity):
         """Return the unit of measurement of this entity, if any."""
 
         if self.tahoma_device.uiclass == "TemperatureSensor":
-            return TEMP_CELSIUS  # TODO Retrieve core:MeasuredValueType to understand if it is Celsius or Kelvin
+            # TODO Retrieve core:MeasuredValueType to understand if it is Celsius or Kelvin
+            return TEMP_CELSIUS
 
         if self.tahoma_device.uiclass == "HumiditySensor":
             return UNIT_PERCENTAGE
@@ -61,7 +72,31 @@ class TahomaSensor(TahomaDevice, Entity):
         if self.tahoma_device.uiclass == "LightSensor":
             return "lx"
 
+        if self.tahoma_device.uiclass == "ElectricitySensor":
+            return POWER_WATT
+
+        if self.tahoma_device.uiclass == "AirSensor":
+            return CONCENTRATION_PARTS_PER_MILLION
+
         return None
+
+    @property
+    def icon(self) -> Optional[str]:
+        """Return the icon to use in the frontend, if any."""
+
+        if self.tahoma_device.uiclass == "AirSensor":
+            return "mdi:periodic-table-co2"
+
+        return None
+
+    @property
+    def device_class(self) -> Optional[str]:
+        """Return the device class of this entity if any."""
+        return (
+            TAHOMA_SENSOR_DEVICE_CLASSES.get(self.tahoma_device.widget)
+            or TAHOMA_SENSOR_DEVICE_CLASSES.get(self.tahoma_device.uiclass)
+            or None
+        )
 
     def update(self):
         """Update the state."""
@@ -86,4 +121,16 @@ class TahomaSensor(TahomaDevice, Entity):
                 )
             )
 
-        _LOGGER.debug("Update %s, value: %d", self._name, self.current_value)
+        if CORE_ELECTRIC_POWER_CONSUMPTION_STATE in self.tahoma_device.active_states:
+            self.current_value = float(
+                "{:.2f}".format(
+                    self.tahoma_device.active_states.get(
+                        CORE_ELECTRIC_POWER_CONSUMPTION_STATE
+                    )
+                )
+            )
+
+        if CORE_CO2_CONCENTRATION_STATE in self.tahoma_device.active_states:
+            self.current_value = int(
+                self.tahoma_device.active_states.get(CORE_CO2_CONCENTRATION_STATE)
+            )