+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added support for Gate devices
+- Added support for AirSensor devices
+- Added support for ElectricitySensor devices
+- Added support for Curtain devices
+- Added support for Generic devices (cover)
+- Added support for SwingingShutter devices
 
 ### Changed